+++ app/src/main/java/net/wildfyre/client/views/markdown/Utils.kt
@@ -6,6 +6,6 @@ import net.wildfyre.client.data.Image
 fun String.prepareForMarkdown(imageUrls: List<Image>?): String =
     replace(Constants.Api.IMAGE_REGEX) {
         val imageNum = it.groups[1]?.value?.toInt() ?: 0
-        val image = imageUrls?.find { img -> img.num == imageNum.toLong() }
+        val image = imageUrls?.find { img -> img.num == imageNum }
         if (imageUrls != null) "![${image?.comment}](${image?.image})" else it.value
     }
